Eartquake/requirements.txt
flask==2.3.3
requests==2.26.0
folium==0.12.1
flask-debugtoolbar==0.13.1
python-dotenv==1.0.0 